backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
psycopg2-binary
pydantic>=2.0
pydantic-settings
python-jose[cryptography]
passlib[argon2]
argon2-cffi
python-dotenv
alembic
structlog
python-json-logger
